+++ R/plot_severity_vs_nphenotypes.R
@@ -22,7 +22,7 @@
 #' out <- plot_severity_vs_nphenotypes(results=results)
 plot_severity_vs_nphenotypes <- function(results,
                                          gpt_annot=HPOExplorer::gpt_annot_codify(),
-                                         cl = KGExplorer::get_ontology("cl", remove_rings=TRUE),
+                                         cl = get_cl(),
                                          q_threshold=0.5,
                                          n_label=3,
                                          size=3,